clearer folder structure; update docs
