+++ README.md
@@ -51,10 +51,12 @@ environment variables (secrets available from project owner):
 ```
 TEST_USER_EMAIL=email@example.com  TEST_USER_PASSWORD=password \
 BROWSERSTACK_USERNAME=foo BROWSERSTACK_KEY=bar \
-SELENIUM_BROWSER=firefox:50:windows \
-SELENIUM_REMOTE_URL=http://hub-cloud.browserstack.com/wd/hub \
-npm run test:acceptance
+npm run test:acceptance -- --reporter [spec|dot|nyan]
 ```
+
+The default reporter is junit, which will not report anything to the cli but to
+a local file `test.results.xml`, hence the above overrides the default with a
+cli reporter.
 
 ## Deploying / working on the charm
